movie-recommender-system/pyproject.toml
[project]
name = "movie-recommender-system-tmdb-dataset-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=6.0.0",
    "nltk>=3.9.4",
    "numpy>=2.4.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.55.0",
]
